Version 0.020
Sets up 2 Grids and contains an abstract class for AIController, which
can be used to create AI schemas. Can't remember if this was in version
0.19, but the game ends when the grid reaches the top of either screen
(including Player 2's grid which currently has nothing controlling it -
so the game automatically ends after a constant time T each time the
program is run).